safetensors

Module slice

Module handling lazy loading via iterating on slices on the original buffer.

Structs

SliceIterator

Iterator used to return the bits of the overall tensor buffer when client asks for a slice of the original tensor.

Enums

InvalidSlice

Error representing invalid slicing attempt

TensorIndexer

Generic structure used to index a slice of the tensor

Traits

IndexOp

Trait used to implement multiple signatures for ease of use of the slicing of a tensor