Trait safetensors::slice::IndexOp

pub trait IndexOp<'data, T> {
    // Required method
    fn slice(
        &'data self,
        index: T
    ) -> Result<SliceIterator<'data>, InvalidSlice>;
}

Trait used to implement multiple signatures for ease of use of the slicing of a tensor

Required Methods

fn slice(&'data self, index: T) -> Result<SliceIterator<'data>, InvalidSlice>

Returns a slicing iterator which are the chunks of data necessary to reconstruct the desired tensor.